+++ src/datastore/sync_methods/defineResource.js
@@ -103,6 +103,7 @@ function defineResource(definition) {
     }
 
     def.getEndpoint = function (attrs, options) {
+      options = DSUtils.deepMixIn({}, options);
       var parent = this.parent;
       var parentKey = this.parentKey;
       var item;